Refactor: move hidden cards into Board
The purpose of this change is to simplify APIs and to let Board itself maintain the status and correctness of the hidden lists, which is easier and less bug-prone than doing it separately.
This also makes Board now a sensitive data structure that can't be passed around to other players (because they can peek into other players' hidden stack).